refactor(web-app): keep display helpers out of the schema module
`types.ts` parses what the server said; how a duration or a missing title
reads on screen is this client's own choice, and the two had grown into
one file that the queue's run time would have pushed over the line cap.

No behaviour changes: the helpers and their tests move as they are.
