+++ packages/web-app/src/views/LibraryView.vue
@@ -15,7 +15,8 @@ import IconClear from '~icons/lucide/x'
 
 import type { LibraryScope } from '../api/ops'
 import type { AlbumEntry, ArtistEntry, GenreEntry, Track } from '../api/types'
-import { coverUrl, formatDuration, QueueMode, trackLabel } from '../api/types'
+import { coverUrl, formatDuration, trackLabel } from '../api/display'
+import { QueueMode } from '../api/types'
 import EmptyState from '../components/EmptyState.vue'
 import PlayingIndicator from '../components/PlayingIndicator.vue'
 import QueueMenu from '../components/QueueMenu.vue'
