+++ packages/web-app/src/views/PlaylistsView.vue
@@ -10,7 +10,8 @@ import IconMusic from '~icons/lucide/music'
 import IconPlay from '~icons/lucide/play'
 import IconSearch from '~icons/lucide/search'
 
-import { coverUrl, formatDuration, QueryField, QueueMode, trackLabel } from '../api/types'
+import { coverUrl, formatDuration, trackLabel } from '../api/display'
+import { QueryField, QueueMode } from '../api/types'
 import EmptyState from '../components/EmptyState.vue'
 import PlayingIndicator from '../components/PlayingIndicator.vue'
 import QueueMenu from '../components/QueueMenu.vue'
