+++ packages/web-app/src/views/QueueView.vue
@@ -8,7 +8,8 @@ import IconSearch from '~icons/lucide/search'
 import IconTrash from '~icons/lucide/trash-2'
 import IconX from '~icons/lucide/x'
 
-import { ShuffleMode, coverUrl, formatDuration, trackLabel } from '../api/types'
+import { coverUrl, formatDuration, trackLabel } from '../api/display'
+import { ShuffleMode } from '../api/types'
 import EmptyState from '../components/EmptyState.vue'
 import PlayingIndicator from '../components/PlayingIndicator.vue'
 import { heardRows } from '../composables/queueHeard'
